+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Added
+
+- `Open game directory` button to File menu
+
 ## 0.9.2 (2024-10-29)
 
 ### Added